requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Data Augmentation
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Metrics and Logging
tensorboard>=2.13.0
